requirements.txt
langchain>=0.2.14
langchain-google-genai>=2.0.0
google-generativeai>=0.7.2
python-dotenv>=1.0.1
